+++ db/migrate/00_create_tests.rb
@@ -1,7 +1,0 @@
-class CreateTests < ActiveRecord::Migration
-  def change
-    create_table :tests do |t|
-      t.string :foo
-    end
-  end
-end