+++ spec/app/models/test.rb
@@ -1,2 +1,8 @@
-class Test < ActiveRecord::Base
+class Test
+  include ActiveModel::Serialization
+  attr_accessor :foo
+
+  def initialize(options)
+    @foo = options[:foo]
+  end
 end